requirements.txt
asttokens==2.4.1
backcall==0.2.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
greenlet==3.0.0
gspread==5.12.0
httplib2==0.22.0
idna==3.4
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.5.0
jupyter_core==5.4.0
matplotlib-inline==0.1.6
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
playwright==1.39.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==11.0.1
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
